requirements.txt
openai
httpx
pillow
datasets
pydantic
pydantic-ai
tqdm
python-dotenv
gradio
numpy
pandas
pyarrow